Chapter 3: Hello, API!
What we have so far with `staticfile` is a handler.
It takes a request and produces a response. All `staticfile`
knows how to do is try to match request paths to paths on disk.

If we want to serve up static files _and_ API requests, we're going
to do a couple more clever things:

- route request starting with '/api' to their specific handlers
- try to serve up everything else as a static file

Thankfully, Iron has pieces for these tasks:

- `router` to match on paths, path parameters, HTTP request verbs
- `mount` to mount multiple handlers at different URL prefixes

( Hack it all together now... )

What we're passing to this router has to implement Handler.
Fortunately there's a blanket implementation for certain types
of functions, so as long as we conform to that interface, we're all good.

( And then to demonstrate that it still works... )

    curl localhost:8888/api/submissions
    curl localhost:8888